jest and webpack and es6
Jest is working with es6, mostly. We have to use require (otherwise we
need to introduce babel) but all the es6 syntax works out of the box so
it's a small price.

webpack is packing and although it's making large bundles and we need to
figure out pruning etc at least it's working

The graphs are working. Very nice. But I think the c172 ceiling is below
20,000 so it's odd that Vx and Vy don't converge. However I do have some
failing tests so ...