higher confidence
I have ratcheted up the confidence with hand calculations that match the
computer's math. The book is definitely a little sloppy in the numerics.

Methodically approached the beCloseTo precision to match expected
significant figures.

Also, critically, mock the book's data plate for the composite
calculations to reduce compounding error in tests.