Lowry had bad math
In Appendix F the equation he gave for finding sigma from pressure
altitude and temperature was simply wrong. I derived a better one
and checked my results with various sources.

Had to update some values in the test vectors, and disable the whole
Appendix F test. But the results for my plane (which were depending on
TAS calculations) are much closer to sane now. Still not quite the climb
rate I expect but it's closer and V_M is reasonable.