Tidying up a bit, perf data from today's flight
I did a test flight with Stephen today and feel fairly good about the
quality of the measurements. Also confirmed experimentally getting
approximately the ROC at nearly gross that the charts would predict
which feels good. It's still saying my ceiling is over 20,000 ft
which it shouldn't be but who knows maybe it is.